requirements.txt
# Qdrant Vector Database
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0

# Machine Learning
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.0.0

# CLI and UI
click>=8.0.0
rich>=13.0.0

# Logging
loguru>=0.7.0

# Observability and Monitoring
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0

# OpenAI API
openai>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.8
langgraph>=0.0.40
beautifulsoup4>=4.12.0
lark>=1.1.0
